pyproject.toml
[tool.poetry]
name = "pymedia"
version = "0.1.0"
description = ""
authors = ["JuanMe16 <juanmena1605@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
pillow = "^10.0.1"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
